refactor(engine): restoring isRoot to skip marks for non-root, adding isConnect on vm
