+++ packages/integration-karma/test/misc/performance-timing/index.spec.js
@@ -177,29 +177,19 @@ if (isUserTimingSupported) {
                                 label: /<x-child> - constructor/,
                             },
                             {
-                                label: /lwc-hydrate/,
-                                children: [
-                                    {
-                                        label: /<x-child> - render/,
-                                    },
-                                    {
-                                        label: /<x-child> - patch/,
-                                    },
-                                ],
+                                label: /<x-child> - render/,
+                            },
+                            {
+                                label: /<x-child> - patch/,
                             },
                             {
                                 label: /<x-child> - constructor/,
                             },
                             {
-                                label: /lwc-hydrate/,
-                                children: [
-                                    {
-                                        label: /<x-child> - render/,
-                                    },
-                                    {
-                                        label: /<x-child> - patch/,
-                                    },
-                                ],
+                                label: /<x-child> - render/,
+                            },
+                            {
+                                label: /<x-child> - patch/,
                             },
                         ],
                     },
